kdw003: update no.9, no.11

+++ nts.uk/uk.at/at.screen/nts.uk.screen.at.app/src/main/java/nts/uk/screen/at/app/dailyperformance/correction/dto/CompensLeaveComDto.java
@@ -20,7 +20,7 @@ public class CompensLeaveComDto {
 
 	private boolean manageTimeOff;
 
-	private Integer compenLeaveRemain;
+	private Double compenLeaveRemain;
 	
 	private Integer timeRemain;
