+++ nts.uk/uk.at/at.screen/nts.uk.screen.at.app/src/main/java/nts/uk/screen/at/app/dailyperformance/correction/dto/ReserveLeaveDto.java
@@ -13,6 +13,6 @@ public class ReserveLeaveDto {
 
 	private boolean manageRemainNumber;
 	
-	private Integer remainNumber; 
+	private Double remainNumber; 
 	
 }
