+++ nts.uk/uk.at/at.screen/nts.uk.screen.at.app/src/main/java/nts/uk/screen/at/app/dailyperformance/correction/dto/SubstVacationDto.java
@@ -18,6 +18,6 @@ public class SubstVacationDto {
 
 	private boolean manageAtr;
 
-	private Integer holidayRemain;
+	private Double holidayRemain;
 	
 }
