+++ nts.uk/uk.at/at.screen/nts.uk.screen.at.app/src/main/java/nts/uk/screen/at/app/dailyperformance/correction/dto/YearHolidaySettingDto.java
@@ -20,7 +20,7 @@ public class YearHolidaySettingDto {
 
 	private boolean manageTimeOff;
 
-	private Integer annualLeaveRemain;
+	private Double annualLeaveRemain;
 	
 	private Integer timeRemain;
